+++ disma/src/domain/category/existing.rs
