+++ disma/src/domain/category/mod.rs
@@ -9,3 +9,6 @@ mod diff;
 
 pub use existing::*;
 mod existing;
+
+pub use commands::*;
+mod commands;
