+++ disma/src/domain/channel/existing.rs
