--- disma/src/domain/diff/mod.rs
@@ -1,5 +1,0 @@
-pub mod base;
-pub mod category;
-pub mod channel;
-pub mod event;
-pub mod roles;
