--- disma/src/domain/entities/mod.rs
@@ -1,6 +1,0 @@
-pub mod category;
-pub mod channel;
-pub mod guild;
-pub mod overwrites;
-pub mod permission;
-pub mod role;
