+++ disma/src/domain/mod.rs
@@ -1,2 +1,8 @@
+pub mod category;
+pub mod channel;
+pub mod commands;
 pub mod diff;
-pub mod entities;
+pub mod guild;
+pub mod overwrites;
+pub mod permission;
+pub mod role;
