+++ disma/src/domain/role/mod.rs
@@ -4,8 +4,11 @@ pub use awaiting::*;
 mod base;
 pub use base::*;
 
+mod existing;
+pub use existing::*;
+
 mod diff;
 pub use diff::*;
 
-mod existing;
-pub use existing::*;
+mod commands;
+pub use commands::*;
