+++ disma/src/infra/discord/dtos/guild.rs
@@ -1,6 +1,6 @@
 use serde::Deserialize;
 
-use crate::domain::entities::guild::GuildSummary;
+use crate::guild::GuildSummary;
 
 #[derive(Deserialize)]
 pub struct GuildResponse {
